src/mcp_switchboard/server.py
"""MCP Switchboard Server - Main entry point."""
import asyncio
import json
from typing import Any

from mcp.server import Server
from mcp.server.stdio import stdio_server
from mcp.types import Tool, TextContent


app = Server("mcp-switchboard")


@app.list_tools()
async def list_tools() -> list[Tool]:
    """List available tools."""
    return [
        Tool(
            name="setup_mcp_servers",
            description="Analyze task and configure required MCP servers",
            inputSchema={
                "type": "object",
                "properties": {
                    "task_description": {
                        "type": "string",
                        "description": "Natural language description of the task",
                    },
                    "agent_type": {
                        "type": "string",
                        "enum": ["cursor", "kiro", "claude_desktop", "claude_code", "custom"],
                        "description": "AI agent platform requesting setup",
                    },
                    "project_path": {
                        "type": "string",
                        "description": "Absolute path to project directory",
                    },
                    "dry_run": {
                        "type": "boolean",
                        "description": "Preview changes without applying",
                        "default": False,
                    },
                },
                "required": ["task_description", "agent_type", "project_path"],
            },
        )
    ]


@app.call_tool()
async def call_tool(name: str, arguments: dict[str, Any]) -> list[TextContent]:
    """Handle tool calls."""
    if name == "setup_mcp_servers":
        result = {
            "success": True,
            "configured_servers": [],
            "ready": True,
            "message": "Basic implementation - full features coming in Phase 2",
        }
        return [TextContent(type="text", text=json.dumps(result, indent=2))]
    
    raise ValueError(f"Unknown tool: {name}")


async def main() -> None:
    """Run the MCP server."""
    async with stdio_server() as (read_stream, write_stream):
        await app.run(
            read_stream,
            write_stream,
            app.create_initialization_options(),
        )


if __name__ == "__main__":
    asyncio.run(main())
