pyproject.toml
[tool.poetry]
name = "mcp-switchboard"
version = "0.1.0"
description = "Intelligent MCP server orchestrator that automates configuration and lifecycle management"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "mcp_switchboard", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.0"
pyyaml = "^6.0"
aiofiles = "^23.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
pytest-cov = "^4.1"
black = "^24.0"
ruff = "^0.3.0"
mypy = "^1.9"

[tool.poetry.scripts]
mcp-switchboard = "mcp_switchboard.server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
